Remove deprecated projections module
