+++ kikuchipy/_util/tests/test_import.py
@@ -134,17 +134,6 @@ class TestImport:
         ):
             _ = kikuchipy.pattern.foo
 
-    def test_import_projections(self):
-        import kikuchipy.projections
-
-        for obj_name in kikuchipy.projections.__all__:
-            getattr(kikuchipy.projections, obj_name)
-        with pytest.raises(
-            AttributeError,
-            match="module 'kikuchipy.projections' has no attribute 'foo'",
-        ):
-            _ = kikuchipy.projections.foo
-
     def test_import_signals(self):
         import kikuchipy.signals
 
@@ -193,7 +182,6 @@ class TestImport:
             "io",
             "load",
             "pattern",
-            "projections",
             "release",
             "set_log_level",
             "signals",
@@ -302,16 +290,6 @@ class TestImport:
             "normalize_intensity",
             "remove_dynamic_background",
             "rescale_intensity",
-        ]
-
-    def test_dir_projections(self):
-        import kikuchipy.projections
-
-        assert dir(kikuchipy.projections) == [
-            "GnomonicProjection",
-            "HesseNormalForm",
-            "LambertProjection",
-            "SphericalProjection",
         ]
 
     def test_dir_signals(self):
